Re-add myau bas fixed code

+++ src/main/java/myau/module/modules/Jesus.java
@@ -1,19 +1,16 @@
-/*
- * Decompiled with CFR 0.152.
- */
 package myau.module.modules;
 
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.util.Locale;
 import myau.module.Module;
 import myau.property.properties.BooleanProperty;
 import myau.property.properties.FloatProperty;
 
-public class Jesus
-extends Module {
+import java.text.DecimalFormat;
+import java.text.DecimalFormatSymbols;
+import java.util.Locale;
+
+public class Jesus extends Module {
     private static final DecimalFormat df = new DecimalFormat("#.##", new DecimalFormatSymbols(Locale.US));
-    public final FloatProperty speed = new FloatProperty("speed", Float.valueOf(2.5f), Float.valueOf(0.0f), Float.valueOf(3.0f));
+    public final FloatProperty speed = new FloatProperty("speed", 2.5F, 0.0F, 3.0F);
     public final BooleanProperty noPush = new BooleanProperty("no-push", true);
     public final BooleanProperty groundOnly = new BooleanProperty("ground-only", true);
 
@@ -26,4 +23,3 @@ extends Module {
         return new String[]{df.format(this.speed.getValue())};
     }
 }
-
