+++ src/main/java/myau/module/modules/Reach.java
@@ -1,12 +1,5 @@
-/*
- * Decompiled with CFR 0.152.
- */
 package myau.module.modules;
 
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.util.Locale;
-import java.util.Random;
 import myau.event.EventTarget;
 import myau.event.types.EventType;
 import myau.events.PickEvent;
@@ -16,12 +9,16 @@ import myau.module.Module;
 import myau.property.properties.FloatProperty;
 import myau.property.properties.PercentProperty;
 
-public class Reach
-extends Module {
+import java.text.DecimalFormat;
+import java.text.DecimalFormatSymbols;
+import java.util.Locale;
+import java.util.Random;
+
+public class Reach extends Module {
     private static final DecimalFormat df = new DecimalFormat("0.0#", new DecimalFormatSymbols(Locale.US));
     private final Random theRandom = new Random();
     private boolean expanding = true;
-    public final FloatProperty range = new FloatProperty("range", Float.valueOf(3.1f), Float.valueOf(3.0f), Float.valueOf(6.0f));
+    public final FloatProperty range = new FloatProperty("range", 3.1F, 3.0F, 6.0F);
     public final PercentProperty chance = new PercentProperty("chance", 100);
 
     public Reach() {
@@ -31,21 +28,21 @@ extends Module {
     @EventTarget
     public void onPick(PickEvent event) {
         if (this.isEnabled() && this.expanding) {
-            event.setRange(((Float)this.range.getValue()).doubleValue());
+            event.setRange(this.range.getValue().doubleValue());
         }
     }
 
     @EventTarget
     public void onRaytrace(RaytraceEvent event) {
         if (this.isEnabled() && this.expanding) {
-            event.setRange(Math.max(event.getRange(), ((Float)this.range.getValue()).doubleValue() + 0.5));
+            event.setRange(Math.max(event.getRange(), this.range.getValue().doubleValue() + 0.5));
         }
     }
 
     @EventTarget
     public void onTick(TickEvent event) {
         if (this.isEnabled() && event.getType() == EventType.PRE) {
-            this.expanding = this.theRandom.nextDouble() <= (double)((Integer)this.chance.getValue()).intValue() / 100.0;
+            this.expanding = this.theRandom.nextDouble() <= (double) this.chance.getValue() / 100.0;
         }
     }
 
@@ -54,4 +51,3 @@ extends Module {
         return new String[]{df.format(this.range.getValue())};
     }
 }
-
